backend/requirements.txt
google-generativeai
python-dotenv
djangorestframework
django-cors-headers
pandas